+++ src/psycopmlutils/feature_describer/feature_describer.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from wasabi import Printer
 
-from psycopmlutils.data_checks.validate_raw_data import save_df_to_pretty_html
+from psycopmlutils.data_checks.utils import save_df_to_pretty_html
 from psycopmlutils.loaders.flattened.local_feature_loaders import load_split_predictors
 from psycopmlutils.utils import generate_feature_colname
